feat: 인벤토리 UI에 핫슬롯에 대한 대응 처리
